Refactored the Renderer#render_string method to take alignment options.
Strings can now be rendered by passing :x, :y, :halign and :valign.
Currently, passing :valign or :halign will mean that :y or :x is
ignored. In the future we will be able to specify a bounding box size to
make the alignment relative to that rather than the screen size.

Added a <title> element.