Use threads to increase performance. This is much better on the Beagle.
One thread now listens on the named pipe for requests. This adds
requests to a queue. The other thread picks up the requests and renders
them.

The performance increase comes from the fact that the rendering thread
is not fast enough to render a request before the next one arrives. So,
the producer thread clears the queue before adding a new request, which
allows Rembrandt to drop frames whenever it needs to (but the final
frame should always be rendered).

Why a queue if it is always cleared, you ask? Simply because it was the
most straightforward to program to get immediate satisfaction!